Call send and estimate using Transaction kit
