+++ src/components/SendNativeToken/index.tsx
@@ -1,13 +1,9 @@
-import React, { useState, KeyboardEvent, useEffect } from 'react';
+import React, {useRef} from 'react';
+import   SendNativeTokenUI   from './sendNativeTokenUI';
 import {
-  EtherspotBatches,
-  EtherspotBatch,
-  EtherspotTransaction,
-  useEtherspotTransactions,
+  EtherspotTransactionKit
 } from "@etherspot/transaction-kit";
 import { ethers } from "ethers";
-import { isValidEthereumAddress } from '../../utils/validation';
-
 
 interface SendNativeTokenProps {
   receiverAddress: string;
@@ -32,134 +28,27 @@ const SendNativeToken = ({
   disableSendOnEnter = false,
   onlyEstimate = false,
 }: SendNativeTokenProps) => {
-
-  const [amount, setAmount] = useState('0.001');
-  const [isValidAddress, setIsValidAddress] = useState(true);
-  const [error, setError] = useState('');
-  const { estimate, send } = useEtherspotTransactions();
   const randomWallet = ethers.Wallet.createRandom();
   const providerWallet = new ethers.Wallet(randomWallet.privateKey);
+  const triggerElementRef = useRef<HTMLButtonElement>(null);
   console.log(providerWallet);
-
-  useEffect(() => {
-    // Add event listener for triggerElement click
-    const triggerElementRef = triggerElement as React.MutableRefObject<HTMLElement> | undefined;
-    if (triggerElementRef && triggerElementRef.current) {
-      triggerElementRef.current.addEventListener('click', estimateAndSend);
-    }
-
-    // Cleanup
-    return () => {
-      if (triggerElementRef && triggerElementRef.current) {
-        triggerElementRef.current.removeEventListener('click', estimateAndSend);
-      }
-    };
-  }, [triggerElement]);
-
-  useEffect(() => {
-    // validate receiver address
-    if(isValidEthereumAddress(receiverAddress)){  
-      setIsValidAddress(true);
-    }else{
-      setIsValidAddress(false);
-    }
-  }, [receiverAddress]);
-
-  // handle on key press logic of sending 
-  const handleEnterPress = (event: KeyboardEvent<HTMLInputElement>) => {
-    if (!disableSendOnEnter && event.key === 'Enter') {
-      // Trigger sending logic here
-      estimateAndSend()
-    }
-  };
-
-  // on change logic
-  const handleAmountChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setAmount(event.target.value);
-  };
-
-  // logic of estimate and send based on props
-  /*const estimateAndSend = async () => {
-    if (isValidAddress) {
-      // Estimate logic
-      const gasEstimation = await transactionKitProps.provider.estimateGas({
-        to: transactionKitProps.to,
-        value: transactionKitProps.amount,
-      });
-  
-      if (!onlyEstimate) {
-        // Send logic
-        try {
-          const transactionHash = await transactionKitProps.provider.sendTransaction({
-            to: transactionKitProps.to,
-            value: transactionKitProps.amount,
-            gasLimit: gasEstimation.toNumber() + 10000, // Adding some buffer to the estimated gas limit
-          });
-  
-          console.log('Transaction sent! Transaction Hash:', transactionHash);
-          // Optionally, you can handle the success of the transaction here
-        } catch (error) {
-          console.error('Error sending transaction:', error);
-          // Optionally, you can handle the failure of the transaction here
-        }
-      }
-  
-      if (debug) {
-        console.log('send');
-      }
-    } else {
-      setError('Receiver address is not a valid blockchain address');
-    }
-  }; */
-
-  const estimateAndSend = async () => {
-    if (isValidAddress) {
-      // Estimate logic
-      const gasEstimation = await estimate();
-  
-      if (!onlyEstimate) {
-       send();
-      }
-  
-      if (debug) {
-        console.log('send');
-      }
-    } else {
-      setError('Receiver address is not a valid blockchain address');
-    }
-  };
-  
-  const transactionKitProps = {
-    provider: providerWallet,
-    to: receiverAddress,
-    chainId: chain,
-    amount: onlyEstimate ? 0 : parseFloat(amount),
-    debug,
-  };
-
   return (
-    <div>
-      <EtherspotBatches>
-        <EtherspotBatch>
-          <EtherspotTransaction to={transactionKitProps.to} value={transactionKitProps.amount}>
-            <div className="mt-2.5">
-              <input
-                type="text"
-                value={amount}
-                onChange={handleAmountChange}
-                onKeyDown={handleEnterPress}
-                placeholder="Enter Amount"
-                className= {unstyled ? '' : className }
-                style={style}
-              />
-              {error ? <p className='text-sm'>{error}</p> : ''}
-            </div>
-          </EtherspotTransaction>
-        </EtherspotBatch>
-      </EtherspotBatches>
-    </div>
+  <EtherspotTransactionKit
+    provider={providerWallet} /* The random wallet we created above */
+    chainId={chain}
+  >
+     <SendNativeTokenUI
+        receiverAddress={receiverAddress}
+        chain={chain} 
+        className= {className}
+        unstyled={unstyled} 
+        debug={debug}
+        disableSendOnEnter={disableSendOnEnter}
+        onlyEstimate={onlyEstimate}
+        style={style}
+        triggerElement={triggerElementRef}
+      />
+     </EtherspotTransactionKit>
   );
-  
 };
-
 export default SendNativeToken;
